requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
astrapy==0.6.2
